Allocations: mv newAllocation to Panel context
This needs to be used by two separate parts of the app

Ideally, we would implement panels with routing, but that feels like a
lot of complexity that we don't need right now. So I'm keeping the basic
functionality the same, but making it easier to open this specific panel
from anywhere in the app.

Supporting changes:

* Create new `NewAllocationWrap` component, since NewAllocation is a
  class component and can't use hooks itself to get data like `budgets`
  directly from `appState`
* This allows simplifying the `setPanel` interface to a point that it
  feels acceptable to move this `newAllocation` function to the shared
  Panel context
* Rename `onNewAllocation` to simply `newAllocation`
