+++ apps/allocations/app/context/Panel.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import { NewAllocation } from '../components/Panel'
 
 const PanelContext = React.createContext()
 
@@ -24,8 +25,18 @@ export function PanelProvider(props) {
     }
   }, [])
 
+  const newAllocation = React.useCallback(budgetId => {
+    setPanel({
+      content: NewAllocation,
+      data: {
+        budgetId,
+        heading: 'New allocation',
+      },
+    })
+  }, [])
+
   const value = React.useMemo(() => {
-    return { panel, panelOpen, setPanel }
+    return { panel, panelOpen, setPanel, newAllocation }
   }, [ panel, panelOpen ])
 
   return <PanelContext.Provider value={value} {...props} />
